requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]
Pillow==11.3.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3